==> ImmuneCellAnalysis/tenX/unTreatedAlone/03_ProjectPrograms.Rmd <==
---
title: "ProjectBroadAnnots.Rmd"
author: "Jenna LaBelle"
date: "10/07/22"
output: html_document
---

## Project program annotations from the full immune cohort to tenX

```{r}
library(Seurat) 

base_dir = "/Users/jlabelle/"
script_dir = "Dropbox (Partners HealthCare)/FilbinLab/data_analysis/Scripts/"
working_dir<- paste0(base_dir, 
                     "Dropbox (Partners HealthCare)/FilbinLab/",
                     "data_analysis/Projects/git_repos/Pedi_Immune/ImmuneCellAnalysis/tenX/unTreatedAlone/")

analysis_dir<- paste0(working_dir, "/analysis/project_programs/")
figure_dir<- paste0(working_dir, "/figures/project_programs/")

if(!dir.exists(figure_dir)){dir.create(figure_dir, recursive = TRUE)}
if(!dir.exists(analysis_dir)){dir.create(analysis_dir, recursive = TRUE)}

preprocessing = paste0(base_dir, script_dir, "single_cell_preprocessing_helper_functions.R")
nmf_helper = paste0(base_dir, script_dir, "NMF_helper_function.R")
source(preprocessing)
source(nmf_helper)

```

## Helper functions
```{r}
## Helper function to identify markers, top markers
my_get_markers<- function(seurat, n_top=10, ident, excludeProgram=NULL){
  seurat@meta.data[["ident"]]<-seurat@meta.data[[ident]]
  seurat<- SetIdent(seurat, value=seurat$ident)
  if(!is.null(excludeProgram)){
    seurat<- subset(seurat, ident != excludeProgram)
  }
  all_markers<- FindAllMarkers(seurat, min.pct = 0.2, logfc.threshold = 0.5)
  markers_filtered<- all_markers[all_markers$p_val_adj<0.05,]
  top_markers<- markers_filtered %>% group_by(cluster) %>% top_n(n=n_top, wt=avg_log2FC) %>% as.data.frame()
  
  return(list(filtered_markers=markers_filtered,
              top_markers=top_markers))
}

## Helper function to run standard seurat projection from reference --> query
my_project_seurat<- function(reference_seurat, query_seurat, reduction="pcaproject",
                             reference_annotation, prediction_colNames=NULL){
  seurat_list<-list(full=reference_seurat, tenx=query_seurat)
  
  ## Transfer anchors and predict
  transferAnchors<-FindTransferAnchors(reference = seurat_list$full, query=seurat_list$tenx,
                                       reduction = reduction)
  
  predictions <- TransferData(anchorset = transferAnchors, 
                              refdata =reference_seurat@meta.data[[reference_annotation]],
                              weight.reduction = reduction)
  
  predictions_final<- predictions$predicted.id; names(predictions_final)<- rownames(predictions)
  
  ## Add to query seurat object
  if(is.null(prediction_colNames)){
      query_seurat<- AddMetaData(query_seurat, predictions, col.name=c("Program_projected", 
                                                         paste0(gsub("prediction.score.", "",
                                                                     colnames(predictions)[2:ncol(predictions)]),
                                                                "_Programscore_projected")))
  }else{
    query_seurat<- AddMetaData(query_seurat, predictions, col.name=prediction_colNames)
  }
  return(query_seurat)

  
                             }


## Helper for barchart by program proportion
my_barchart_programProp<- function(seurat, grouping_variable="sample", coloring_variable,
                                   colors){
  df<- as.data.frame(table(seurat@meta.data[[grouping_variable]], seurat@meta.data[[coloring_variable]]))
  colnames(df)<- c("Sample", "Program", "NCells")
  df<- df %>% group_by(Sample) %>% mutate(perCells=NCells/sum(NCells)) %>% as.data.frame()
  ggplot(df, aes(x=Sample, y=perCells, fill=Program))+
    geom_bar(stat="identity")+
    scale_fill_manual(values=colors[names(colors) %in% df$Program])+
    theme_classic()
                                   }
## Order features for plotting
## input should be of the format from seurat's FindAllMarkers
my_orderFeatures_forDotplot<- function(markers){
      markers$cluster<- as.character(markers$cluster)
      markers$cluster<- factor(markers$cluster, 
                                            levels=unique(markers$cluster)[order(unique(markers$cluster))])
      markers$gene<- factor(markers$gene, levels=markers$gene[order(markers$cluster)])
      markers<- markers[order(markers$gene),]
      return(markers)
}



## Helper function for dotplot of markers of interest
my_dotplot<- function(seurat, grouping_variable, features){
  seurat@meta.data[["tmp"]]<- seurat@meta.data[[grouping_variable]]
  DotPlot(seurat, group.by = "tmp", features=features)+
    coord_flip()+
    theme(axis.text.x = element_text(angle=45, hjust=1))+
    scale_color_gradient2(low="blue", mid="white", high="red")
}
  
## Helper for plotting histogram of max score
## Helps to determine reasonable "Score too low" threshold
my_histMaxScores<- function(seurat, grouping_variable,score_variable, colors){
  seurat@meta.data[["grouping_variable"]]<- seurat@meta.data[[grouping_variable]]
  seurat@meta.data[["score_variable"]]<- seurat@meta.data[[score_variable]]
  cowplot::plot_grid(plotlist=lapply(unique(seurat$grouping_variable), function(x){
    tmp<- subset(seurat, grouping_variable==x)
    ggplot(tmp@meta.data, aes(x=score_variable, fill=grouping_variable))+
      geom_histogram()+
      scale_fill_manual(values=colors[x])+
      theme_classic()+
      theme(legend.position = "none")+
      ggtitle(x)
  }))
}


```

## Load seurat objects: tenx (processed in 02b_ProjectBroad.DetailedAnnots.Rmd) and full immune cohort (both myeloid and t cells)
```{r}
full_cohort_dir<- "../../"
tenx_data_dir<- paste0(working_dir, "/analysis/project_broad.detailed_annot/")

## tenx seurat objects
tenx_myeloid<- readRDS(paste0(tenx_data_dir, "myeloid_tenx_noDC.B_seurat.Rds"))
tenx_cd4<- readRDS(paste0(tenx_data_dir, "cd4_tenx_seurat.Rds"))
tenx_cd8<- readRDS(paste0(tenx_data_dir, "cd8_tenx_seurat.Rds"))

## Full cohort seurat objects
full_myeloid_seurat<- readRDS(paste0(full_cohort_dir,
                                     "01_Preprocessing/02_Counts_Preprocess/analysis/detailed_annotation/",
                                     "pedOnly_nomuv63/",
                                     "myeloid_seurat_noDC.B_harmony.Rds"))
full_cd4_seurat<- readRDS(paste0(full_cohort_dir,
                                   "02a_Tcells/02_Ped.Adult/03_Visualize.Reannotate_programs/analysis/Reannotate_programs/",
                                   "rank8_sharedByunion/",
                                   "ped_seurat_cd4_harmony.Rds"))
full_cd8_seurat<- readRDS(paste0(full_cohort_dir,
                                   "02a_Tcells/02_Ped.Adult/03_Visualize.Reannotate_programs/analysis/Reannotate_programs/",
                                   "rank8_sharedByunion/",
                                   "ped_seurat_cd8_harmony.Rds"))

## Colors 
tcell_colors<- readRDS(paste0(full_cohort_dir, "/plot_colors/Tcell.programs_rank8sharedByunion_reAnnot.Rds"))
myeloid_colors<- readRDS(paste0(full_cohort_dir, "/plot_colors/Myeloid_program_colors.Rds"))
myeloid_colors<- myeloid_colors$rank6

## Full cohort marker genes
tcell_markers<- readRDS(paste0(full_cohort_dir, 
                             "02a_Tcells/02_Ped.Adult/01_compare_NMF/analysis/newMarkers_deNovoPed.AdultTcell/", 
                             "new_markers_NMFrank8_bySeurat_sharedByunion_reAnnot.Rds"))
myeloid_markers<- read.csv(paste0(full_cohort_dir,
                               "02b_Myeloid/01a_NMF/analysis/de.novoNMF/pedOnly_nomuv63/", 
                               "top_DE_genes_10Xannotated_NMF_rank6.csv"))

  ## Top 3 markers per program
  top_tcell_markers<- tcell_markers %>% group_by(cluster) %>% top_n(n=3, wt=avg_log2FC) %>% as.data.frame()
  top_myeloid_markers<- myeloid_markers %>% group_by(cluster) %>% top_n(n=3, wt=avg_log2FC) %>% as.data.frame()

```


## Projection of cd4 programs to tenx cd4
```{r}
## Project full to tenx
tenx_cd4<- my_project_seurat(full_cd4_seurat, tenx_cd4, reference_annotation = "NMF_rank8_annot_mergedByunion_reAnnot")


## plot
DimPlot(tenx_cd4, group.by="Program_projected", pt.size = 1, label=FALSE, label.size = 6)+ 
  scale_color_manual(values=tcell_colors[names(tcell_colors) %in% tenx_cd4$Program_projected])
ggsave(paste0(figure_dir, "UMAP_CD4_program_projectedFromFull.png"), width=6, height=6)

## Proportion of programs by sample
my_barchart_programProp(tenx_cd4, coloring_variable = "Program_projected", colors=tcell_colors)
ggsave(paste0(figure_dir, "Breakdown_CD4_programBySample.png"), width=4, height=4)

## Expression of expected (same as full cohort) genes in predicted programs
my_dotplot(tenx_cd4, "Program_projected", unique(my_orderFeatures_forDotplot(top_tcell_markers)$gene))
ggsave(paste0(figure_dir, "DotPlot_cd4_topMarkersForFullCohort.png"), width=6, height=8)

## Identify markers for project programs- match expectations?
cd4_markers<- my_get_markers(tenx_cd4, n_top=5, ident="Program_projected")

  ## Confirm expression of expected markers within predicted programs
  my_dotplot(tenx_cd4, "Program_projected", unique(my_orderFeatures_forDotplot(cd4_markers$top_markers)$gene))
  ggsave(paste0(figure_dir, "DotPlot_CD4_topMarkersForReidentifiedPredictedPrograms.png"), width=6, height=8)
  
## Histogram of max program scores for each program
## Helps to determine reasonable "Score too low" threshold
my_histMaxScores(tenx_cd4, "Program_projected", "max_Programscore_projected", tcell_colors)  
ggsave(paste0(figure_dir, "Histogram_CD4_maxProgramScore.png"), width=12, height=9)

## Grey out cells with low confidence score
low_score<- 0.4
tenx_cd4$Program_projected_lowScore<-ifelse(tenx_cd4$max_Programscore_projected<low_score, 
                                           "ScoreTooLow",
                                           tenx_cd4$Program_projected)
DimPlot(tenx_cd4, group.by = "Program_projected_lowScore")+
  scale_color_manual(values=c(tcell_colors[names(tcell_colors) %in% tenx_cd4$Program_projected],
                              ScoreTooLow="grey"))
ggsave(paste0(figure_dir, "UMAP_CD4_maxProgram_scoreTooLow", low_score, ".png"), width=6.5, height=5)

  ## Expression of expected (same as full cohort) genes in predicted programs
  my_dotplot(subset(tenx_cd4, Program_projected_lowScore != "ScoreTooLow"), 
             "Program_projected_lowScore", unique(my_orderFeatures_forDotplot(top_tcell_markers)$gene))
  ggsave(paste0(figure_dir, "DotPlot_CD4_topMarkersForFullCohort_scoreTooLow", low_score, ".png"), width=6, height=8)

  ## Identify markers for project programs- match expectations?
  cd4_markers_lowScore<- my_get_markers(tenx_cd4, n_top=5, ident="Program_projected_lowScore", excludeProgram = "ScoreTooLow")
  
  ## Confirm expression of expected markers within predicted programs
  my_dotplot(tenx_cd4, "Program_projected_lowScore", unique(my_orderFeatures_forDotplot(cd4_markers_lowScore$top_markers)$gene))
  ggsave(paste0(figure_dir, "DotPlot_CD4_topMarkersForReidentifiedPredictedPrograms_scoreTooLow", low_score, ".png"), width=6, height=8)

saveRDS(tenx_cd4, file=paste0(tenx_data_dir, "cd4_tenx_seurat.Rds"))
```

## Projection of cd8 programs to tenx cd8
```{r}
tenx_cd8<- my_project_seurat(full_cd8_seurat, tenx_cd8, reference_annotation = "NMF_rank8_annot_mergedByunion_reAnnot")

## plot
DimPlot(tenx_cd8, group.by="Program_projected", pt.size = 1, label=FALSE, label.size = 6)+ 
  scale_color_manual(values=tcell_colors[names(tcell_colors) %in% tenx_cd8$Program_projected])
ggsave(paste0(figure_dir, "UMAP_CD8_program_projectedFromFull.png"), width=6, height=6)

## Proportion of programs by sample
my_barchart_programProp(tenx_cd8, coloring_variable = "Program_projected", colors=tcell_colors)
ggsave(paste0(figure_dir, "Breakdown_cd8_programBySample.png"), width=4, height=4)

## Expression of expected (same as full cohort) genes in predicted programs
my_dotplot(tenx_cd8, "Program_projected", unique(my_orderFeatures_forDotplot(top_tcell_markers)$gene))
ggsave(paste0(figure_dir, "DotPlot_cd8_topMarkersForFullCohort.png"), width=6, height=8)

## Identify markers for project programs- match expectations?
cd8_markers<- my_get_markers(tenx_cd8, n_top=5, ident="Program_projected")

  ## Expression of top markers for each predicted program
  my_dotplot(tenx_cd8, "Program_projected", unique(my_orderFeatures_forDotplot(cd8_markers$top_markers)$gene))
  ggsave(paste0(figure_dir, "DotPlot_cd8_topMarkersForReidentifiedPredictedPrograms.png"), width=6, height=8)

  
## Histogram of max program scores for each program
## Helps to determine reasonable "Score too low" threshold
my_histMaxScores(tenx_cd8, "Program_projected", "max_Programscore_projected", tcell_colors)  
ggsave(paste0(figure_dir, "Histogram_CD8_maxProgramScore.png"), width=12, height=9)

## Grey out cells with low confidence score
low_score<- 0.4
tenx_cd8$Program_projected_lowScore<-ifelse(tenx_cd8$max_Programscore_projected<low_score, 
                                           "ScoreTooLow",
                                           tenx_cd8$Program_projected)
DimPlot(tenx_cd8, group.by = "Program_projected_lowScore")+
  scale_color_manual(values=c(tcell_colors[names(tcell_colors) %in% tenx_cd8$Program_projected],
                              ScoreTooLow="grey"))
ggsave(paste0(figure_dir, "UMAP_CD8_maxProgram_scoreTooLow", low_score, ".png"), width=6.5, height=5)

  ## Expression of expected (same as full cohort) genes in predicted programs
  my_dotplot(subset(tenx_cd8, Program_projected_lowScore != "ScoreTooLow"), 
             "Program_projected_lowScore", unique(my_orderFeatures_forDotplot(top_tcell_markers)$gene))
  ggsave(paste0(figure_dir, "DotPlot_cd8_topMarkersForFullCohort_scoreTooLow", low_score, ".png"), width=6, height=8)

  ## Identify markers for project programs- match expectations?
  cd8_markers_lowScore<- my_get_markers(tenx_cd8, n_top=5, ident="Program_projected_lowScore", excludeProgram = "ScoreTooLow")
  
  ## Confirm expression of expected markers within predicted programs
  my_dotplot(tenx_cd8, "Program_projected_lowScore", unique(my_orderFeatures_forDotplot(cd8_markers_lowScore$top_markers)$gene))
  ggsave(paste0(figure_dir, "DotPlot_cd8_topMarkersForReidentifiedPredictedPrograms_scoreTooLow", low_score, ".png"), width=6, height=8)

saveRDS(tenx_cd8, file=paste0(tenx_data_dir, "cd8_tenx_seurat.Rds"))
```

## Projection of myeloid programs to tenx myeloid
```{r}
tenx_myeloid<- my_project_seurat(full_myeloid_seurat, tenx_myeloid, reference_annotation = "NMF_rank6_annotByAdult10X")

## plot
DimPlot(tenx_myeloid, group.by="Program_projected", pt.size = 1, label=FALSE, label.size = 6)+ 
  scale_color_manual(values=myeloid_colors[names(myeloid_colors) %in% tenx_myeloid$Program_projected])
ggsave(paste0(figure_dir, "UMAP_myeloid_program_projectedFromFull.png"), width=6, height=6)

## Proportion of programs by sample
my_barchart_programProp(tenx_myeloid, coloring_variable = "Program_projected", colors=myeloid_colors)
ggsave(paste0(figure_dir, "Breakdown_myeloid_programBySample.png"), width=4, height=4)

## Expression of expected (same as full cohort) genes in predicted programs
my_dotplot(tenx_myeloid, "Program_projected", unique(my_orderFeatures_forDotplot(top_myeloid_markers)$gene))
ggsave(paste0(figure_dir, "DotPlot_myeloid_topMarkersForFullCohort.png"), width=6, height=8)

## Identify markers for project programs- match expectations?
myeloid_markers<- my_get_markers(tenx_myeloid, n_top=5, ident="Program_projected")

  ## Expression of top markers for each predicted program
  my_dotplot(tenx_myeloid, "Program_projected", unique(my_orderFeatures_forDotplot(myeloid_markers$top_markers)$gene))
  ggsave(paste0(figure_dir, "DotPlot_myeloid_topMarkersForReidentifiedPredictedPrograms.png"), width=6, height=8)

  
## Histogram of max program scores for each program
## Helps to determine reasonable "Score too low" threshold
my_histMaxScores(tenx_myeloid, "Program_projected", "max_Programscore_projected", myeloid_colors)  
ggsave(paste0(figure_dir, "Histogram_myeloid_maxProgramScore.png"), width=12, height=9)

## Grey out cells with low confidence score
low_score<- 0.4
tenx_myeloid$Program_projected_lowScore<-ifelse(tenx_myeloid$max_Programscore_projected<low_score, 
                                           "ScoreTooLow",
                                           tenx_myeloid$Program_projected)
DimPlot(tenx_myeloid, group.by = "Program_projected_lowScore")+
  scale_color_manual(values=c(myeloid_colors[names(myeloid_colors) %in% tenx_myeloid$Program_projected],
                              ScoreTooLow="grey"))
ggsave(paste0(figure_dir, "UMAP_myeloid_maxProgram_scoreTooLow", low_score, ".png"), width=6.5, height=5)

  ## Expression of expected (same as full cohort) genes in predicted programs
  my_dotplot(subset(tenx_myeloid, Program_projected_lowScore != "ScoreTooLow"), 
             "Program_projected_lowScore", unique(my_orderFeatures_forDotplot(top_myeloid_markers)$gene))
  ggsave(paste0(figure_dir, "DotPlot_myeloid_topMarkersForFullCohort_scoreTooLow", low_score, ".png"), width=6, height=8)

  ## Identify markers for project programs- match expectations?
  myeloid_markers_lowScore<- my_get_markers(tenx_myeloid, n_top=5, ident="Program_projected_lowScore", excludeProgram = "ScoreTooLow")
  
  ## Confirm expression of expected markers within predicted programs
  my_dotplot(tenx_myeloid, "Program_projected_lowScore", unique(my_orderFeatures_forDotplot(myeloid_markers_lowScore$top_markers)$gene))
  ggsave(paste0(figure_dir, "DotPlot_myeloid_topMarkersForReidentifiedPredictedPrograms_scoreTooLow", low_score, ".png"), width=6, height=8)

saveRDS(tenx_myeloid, file=paste0(tenx_data_dir, "myeloid_tenx_seurat.Rds"))
```
