==> ImmuneCellAnalysis/tenX/unTreatedAlone/04_ManualProgramValidation.Rmd <==
---
title: "ProjectBroadAnnots.Rmd"
author: "Jenna LaBelle"
date: "10/07/22"
output: html_document
---

## Annotate graph-based clusters for cd4/cd8 T cells. Used to validate NMF-based programs in full cohort

```{r}
library(Seurat) 

base_dir = "/Users/jlabelle/"
script_dir = "Dropbox (Partners HealthCare)/FilbinLab/data_analysis/Scripts/"
working_dir<- paste0(base_dir, 
                     "Dropbox (Partners HealthCare)/FilbinLab/",
                     "data_analysis/Projects/git_repos/Pedi_Immune/ImmuneCellAnalysis/tenX/unTreatedAlone/")

analysis_dir<- paste0(working_dir, "/analysis/validate_programs/")
figure_dir<- paste0(working_dir, "/figures/validate_programs/")

if(!dir.exists(figure_dir)){dir.create(figure_dir, recursive = TRUE)}
if(!dir.exists(analysis_dir)){dir.create(analysis_dir, recursive = TRUE)}

preprocessing = paste0(base_dir, script_dir, "single_cell_preprocessing_helper_functions.R")
nmf_helper = paste0(base_dir, script_dir, "NMF_helper_function.R")
source(preprocessing)
source(nmf_helper)

```


## Load tenx (processed in 02b_ProjectBroad.DetailedAnnots.Rmd) 
```{r}
tenx_data_dir<- paste0(working_dir, "/analysis/project_broad.detailed_annot/")
full_cohort_dir<- "../../"

## tenx seurat objects
tenx_cd4<- readRDS(paste0(tenx_data_dir, "cd4_tenx_seurat.Rds"))
tenx_cd8<- readRDS(paste0(tenx_data_dir, "cd8_tenx_seurat.Rds"))


## Colors 
tcell_colors<- readRDS(paste0(full_cohort_dir, "/plot_colors/Tcell.programs_rank8sharedByunion_reAnnot.Rds"))

## Full cohort marker genes
tcell_markers<- readRDS(paste0(full_cohort_dir, 
                             "02a_Tcells/02_Ped.Adult/01_compare_NMF/analysis/newMarkers_deNovoPed.AdultTcell/", 
                             "new_markers_NMFrank8_bySeurat_sharedByunion_reAnnot.Rds"))

  ## Top 3 markers per program
  top_tcell_markers<- tcell_markers %>% group_by(cluster) %>% top_n(n=3, wt=avg_log2FC) %>% as.data.frame()

```

## CD8
```{r}
## Change cluster resolution
tenx_cd8<-FindClusters(tenx_cd8, resolution = 2)
DimPlot(tenx_cd8, label=TRUE)+NoLegend()
ggsave(paste0(figure_dir, "UMAP_CD8_clusters_res2.png"), width=5, height=5)

## Find marker genes for clusters
markers<- FindAllMarkers(tenx_cd8)
filtered_markers<- markers[markers$p_val_adj<0.05 & markers$avg_log2FC>0,]
top_markers<- filtered_markers %>% group_by(cluster) %>% top_n(n=10, wt=avg_log2FC) %>% as.data.frame()

## Heatmap of top markers
tenx_cd8_rescaled<- ScaleData(tenx_cd8, do.scale=FALSE, do.center=TRUE, features=rownames(tenx_cd8))
DoHeatmap(tenx_cd8_rescaled, features=top_markers$gene)+
  scale_fill_gradient2(low="blue", mid="white", high="red")
ggsave(paste0(figure_dir, "Heatmap_CD8_topClusterMarkers.png"), width=5, height=6)


## Score CD8 for cluster markers + full cohort markers
cm_list<- NormCenter(tenx_cd8@assays$RNA@counts)
cm_mean<- rowMeans(log2(cm_list$raw_data + 1))

  ## Split marker dfs to list
  cluster_marker_list<- split(filtered_markers, f=filtered_markers$cluster)
  cluster_marker_list<- lapply(cluster_marker_list, function(x){x$gene})
  full_marker_list<- split(tcell_markers, f=tcell_markers$cluster)
  full_marker_list<- lapply(full_marker_list, function(x){x$gene})
  full_marker_list<- full_marker_list[!names(full_marker_list) %in% c("Inhibitory", "Treg")]
  
  ## Score tenx cd8 for both cluster markers + full cohort nmf-based markers
  scores<- as.data.frame(t(scoreNmfGenes(cm_list$center_data, cm_mean, c(cluster_marker_list, full_marker_list))))

## Correlate scores for cluster markers/full cohort markers
  scores_factor_hc = clusterNmfFactors(scores)
  scores_factor_cor = scores_factor_hc$cor_coef[scores_factor_hc$hc_obj$order, scores_factor_hc$hc_obj$order]
  
  ## Heatmap of correlations
  hm_colors = rev((brewer.pal(n=9, name="RdBu")))
  hm_colors = colorRampPalette(colors = hm_colors)
  pheatmap(scores_factor_cor, color = hm_colors(100), 
           cluster_rows = F, cluster_cols = F, 
           annotation_names_row = F, annotation_names_col =T,
           show_rownames = T, show_colnames = F,
           filename = paste0(figure_dir, "Cor_CD8_clusterMarkers_vs_fullCohortMarkers.pdf"),
           width = 8, height = 4)
  
  ## Dendrogram of hierarchical clustering 
  jpeg(filename=paste0(figure_dir, "HC_CD8_clusterMarkers_vs_fullCohortMarkers.pdf"), width = 400, height = 600)
  plot(scores_factor_hc$hc_obj)
  dev.off()

## Annotate clusters based on correlation/marker genes
clusters<- as.character(unique(tenx_cd8$seurat_clusters))
names(clusters)<- c("IFN.Response", "Cycling", "Cytotoxic.NKreceptor", "Predysfunctional", "Memory")
tenx_cd8$seurat_clusters_annot<- as.character(plyr::mapvalues(tenx_cd8$seurat_clusters,
                                                              clusters, names(clusters)))

## Plot: UMAP, heatmap
DimPlot(tenx_cd8, group.by = "seurat_clusters_annot", label=TRUE)+ NoLegend()+
  scale_color_manual(values=tcell_colors)
ggsave(paste0(figure_dir, "UMAP_CD8_clusters_annot.pdf"),device="pdf", width=5, height=5)


tenx_cd8_rescaled<- ScaleData(tenx_cd8, do.scale=FALSE, do.center=TRUE, features=rownames(tenx_cd8))
tenx_cd8_rescaled$seurat_clusters_annot<- factor(tenx_cd8_rescaled$seurat_clusters_annot, levels=c("Memory", "Cycling",
                                                                                                   "Cytotoxic.NKreceptor",
                                                                                                   "IFN.Response", "Predysfunctional"))
DoHeatmap(tenx_cd8_rescaled, features=top_markers$gene, group.by = "seurat_clusters_annot", group.colors = tcell_colors)+
  scale_fill_gradient2(low="blue", mid="white", high="red")
ggsave(paste0(figure_dir, "Heatmap_CD8_topClusterMarkers_withClusterAnnot.png"), device="png", width=5, height=7)

write.csv(filtered_markers, file=paste0(analysis_dir, "CD8_cluster_markers_filtered.csv"))
saveRDS(tenx_cd8, file=paste0(tenx_data_dir, "cd8_tenx_seurat.Rds"))
```

## CD4
```{r}
## Change cluster resolution
tenx_cd4<-FindClusters(tenx_cd4, resolution = .4)
DimPlot(tenx_cd4, label=TRUE)+NoLegend()
ggsave(paste0(figure_dir, "UMAP_CD4_clusters_res0.4.png"), width=5, height=5)

## Find marker genes for clusters
markers<- FindAllMarkers(tenx_cd4)
filtered_markers<- markers[markers$p_val_adj<0.05 & markers$avg_log2FC>0,]
top_markers<- filtered_markers %>% group_by(cluster) %>% top_n(n=10, wt=avg_log2FC) %>% as.data.frame()

## Heatmap of top markers
tenx_cd4_rescaled<- ScaleData(tenx_cd4, do.scale=FALSE, do.center=TRUE, features=rownames(tenx_cd4))
DoHeatmap(tenx_cd4_rescaled, features=top_markers$gene)+
  scale_fill_gradient2(low="blue", mid="white", high="red")
ggsave(paste0(figure_dir, "Heatmap_CD4_topClusterMarkers.png"), width=5, height=6)


## Score cd4 for cluster markers + full cohort markers
cm_list<- NormCenter(tenx_cd4@assays$RNA@counts)
cm_mean<- rowMeans(log2(cm_list$raw_data + 1))

  ## Split marker dfs to list
  cluster_marker_list<- split(filtered_markers, f=filtered_markers$cluster)
  cluster_marker_list<- lapply(cluster_marker_list, function(x){x$gene})
  full_marker_list<- split(tcell_markers, f=tcell_markers$cluster)
  full_marker_list<- lapply(full_marker_list, function(x){x$gene})
  full_marker_list<- full_marker_list[!names(full_marker_list) %in% c("Predysfunctional", "Cytotoxic.NKreceptor")]
  
  ## Score tenx cd4 for both cluster markers + full cohort nmf-based markers
  scores<- as.data.frame(t(scoreNmfGenes(cm_list$center_data, cm_mean, c(cluster_marker_list, full_marker_list))))

## Correlate scores for cluster markers/full cohort markers
  scores_factor_hc = clusterNmfFactors(scores)
  scores_factor_cor = scores_factor_hc$cor_coef[scores_factor_hc$hc_obj$order, scores_factor_hc$hc_obj$order]
  
  ## Heatmap of correlations
  hm_colors = rev((brewer.pal(n=9, name="RdBu")))
  hm_colors = colorRampPalette(colors = hm_colors)
  pheatmap(scores_factor_cor, color = hm_colors(100), 
           cluster_rows = F, cluster_cols = F, 
           annotation_names_row = F, annotation_names_col =T,
           show_rownames = T, show_colnames = F,
           filename = paste0(figure_dir, "Cor_CD4_clusterMarkers_vs_fullCohortMarkers.pdf"),
           width = 8, height = 4)
  
  ## Dendrogram of hierarchical clustering 
  jpeg(filename=paste0(figure_dir, "HC_CD4_clusterMarkers_vs_fullCohortMarkers.pdf"), width = 400, height = 600)
  plot(scores_factor_hc$hc_obj)
  dev.off()

## Annotate clusters based on correlation/marker genes
clusters<- as.character(unique(tenx_cd4$seurat_clusters))
names(clusters)<- c( "Memory", "Inhibitory", "Cycling")
tenx_cd4$seurat_clusters_annot<- as.character(plyr::mapvalues(tenx_cd4$seurat_clusters,
                                                              clusters, names(clusters)))

## Plot: UMAP, heatmap
DimPlot(tenx_cd4, group.by = "seurat_clusters_annot", label=TRUE)+ NoLegend()+
  scale_color_manual(values=tcell_colors)
ggsave(paste0(figure_dir, "UMAP_CD4_clusters_annot.pdf"),device="pdf", width=5, height=5)


tenx_cd4_rescaled<- ScaleData(tenx_cd4, do.scale=FALSE, do.center=TRUE, features=rownames(tenx_cd4))
tenx_cd4_rescaled$seurat_clusters_annot<- factor(tenx_cd4_rescaled$seurat_clusters_annot, levels=c("Memory",
                                                                                                   "Inhibitory",
                                                                                                   "Cycling"))
DoHeatmap(tenx_cd4_rescaled, features=top_markers$gene, group.by = "seurat_clusters_annot", group.colors = tcell_colors)+
  scale_fill_gradient2(low="blue", mid="white", high="red")
ggsave(paste0(figure_dir, "Heatmap_CD4_topClusterMarkers_withClusterAnnot.pdf"), device="pdf", width=5, height=7)

write.csv(filtered_markers, file=paste0(analysis_dir, "CD4_cluster_markers_filtered.csv"))
saveRDS(tenx_cd4, file=paste0(tenx_data_dir, "cd4_tenx_seurat.Rds"))
```
