==> ImmuneCellAnalysis/tenX/unTreatedAlone/05_GOI_ExpressionPatterns.Rmd <==
---
title: "R Notebook"
output: html_document
---

## Plot GOI across tenX sample

```{r}
knitr::opts_chunk$set(echo = TRUE)

base_dir = "/Users/jlabelle/"
working_dir<- paste0(base_dir, 
                     "Dropbox (Partners HealthCare)/FilbinLab/",
                     "data_analysis/Projects/git_repos/Pedi_Immune/ImmuneCellAnalysis/tenX/unTreatedAlone/")
script_dir = "Dropbox (Partners HealthCare)/FilbinLab/data_analysis/Scripts/"


preprocessing = paste0(base_dir, script_dir, "single_cell_preprocessing_helper_functions.R")
nmf_helper = paste0(base_dir, script_dir, "NMF_helper_function.R")
source(preprocessing)
source(nmf_helper)

library(Matrix)

analysis_dir<- "analysis/GOI_Expression/"
figure_dir<- "figures/GOI_Expression/"
data_dir<- "data/"
if(!dir.exists(figure_dir)){dir.create(figure_dir)};if(!dir.exists(analysis_dir)){dir.create(analysis_dir)}

marker_dir<- paste0(base_dir, "Dropbox (Partners HealthCare)/FilbinLab/data_analysis/Shared/Marker_genes/")
hkgenes <- read.table(paste0(marker_dir, "tirosh_house_keeping.txt"), skip = 2)
hkgenes <- as.vector(hkgenes$V1)


```


## Read in seurat
```{r}
full_cohort_dir<- "../../"

tcell_seurat<- readRDS("analysis/project_broad.detailed_annot/tcell_tenx_seurat.Rds")
cd8_seurat<- readRDS("analysis/project_broad.detailed_annot/cd8_tenx_seurat.Rds")
cd4_seurat<- readRDS("analysis/project_broad.detailed_annot/cd4_tenx_seurat.Rds")

## Set cluster resolution to use (based on 04_ManualProgramValidation)
cd8_seurat<- FindClusters(cd8_seurat, resolution = 2)
cd4_seurat<- FindClusters(cd4_seurat, resolution=0.4)

## T cell colors
tcell_colors<- readRDS(paste0(full_cohort_dir, "plot_colors/Tcell.programs_rank8sharedByunion_reAnnot.Rds"))

## Read in t cell markers
rank_use<-8; sharedBy<- "union"
program_markers<- readRDS(paste0(full_cohort_dir, "02a_Tcells/", 
                                 "02_Ped.Adult/01_compare_NMF/analysis/newMarkers_deNovoPed.AdultTcell/",
                                 "new_markers_NMFrank", rank_use, "_bySeurat_sharedBy", sharedBy, "_reAnnot.Rds"))
```

## Expression of GOI in CD4/CD8 T cells
```{r}
## Dotplot
DotPlot(tcell_seurat, group.by = "detailed_annot_projected_manual", features=c("SELPLG", "PTGER2"), scale=TRUE)+
  scale_color_gradient2(low="blue", mid="white", high="red")+
  ylab("") + xlab("")+
  theme(axis.text.x = element_text(angle=45, hjust=1, face="bold"),
        axis.text.y = element_text(face="bold"))
ggsave(paste0(figure_dir, "DotPlot_SELPLG.PTGER2_CD4.v.CD8.png"), width=4, height=4)

## Featureplot
tcell_rescaled<- ScaleData(tcell_seurat, do.scale = FALSE, do.center=TRUE, features=rownames(tcell_seurat))
cowplot::plot_grid(plotlist=lapply(c("SELPLG", "PTGER2"), function(x){
  FeaturePlot(tcell_rescaled, features=x,slot="scale")+
  scale_color_gradient2(low="blue", mid="white", high="red")+
  ylab("") + xlab("")+
  theme(axis.text.x = element_text(angle=45, hjust=1, face="bold"),
        axis.text.y = element_text(face="bold"))
}))

ggsave(paste0(figure_dir, "UMAP_SELPLG.PTGER2_Tcells.png"), width=8, height=4)
```



## CD8: plot PTGER2/SELPLG expression by cluster
```{r}
## FeaturePlot
p1<-FeaturePlot(cd8_seurat, features="PTGER2")+scale_color_gradient(low="grey80", high="red")
p2<-DimPlot(cd8_seurat,label=TRUE)+NoLegend()
p3<- VlnPlot(cd8_seurat, features="PTGER2")
p1+p2+p3
ggsave(paste0(figure_dir, "CD8_PTGER2.expression.png"),width=6, height=15)

## by cluster-based annotation (04_ManualProgramValidation.Rmd)
cowplot::plot_grid(plotlist=lapply(c("PTGER2", "SELPLG"), function(x){
  VlnPlot(cd8_seurat,group.by = "seurat_clusters_annot", features=x, sort=FALSE)+NoLegend()+
    scale_fill_manual(values=tcell_colors)
}))
ggsave(paste0(figure_dir, "Vln_CD8_byClusterAnnotated_PTGER2.SELPLG.pdf"), device="pdf", width=8, height=5)

## by projection from full cohort (03_ProjectPrograms.Rmd)
cowplot::plot_grid(plotlist=lapply(c("PTGER2", "SELPLG"), function(x){
  VlnPlot(cd8_seurat,group.by = "Program_projected", features=x, sort=FALSE)+NoLegend()+
    scale_fill_manual(values=tcell_colors)
}))
ggsave(paste0(figure_dir, "Vln_CD8_byProjectedProgram_PTGER2.SELPLG.pdf"), device="pdf", width=8, height=5)

```


## CD4: plot PTGER2/SELPLG expression by cluster
```{r}
## FeaturePlot
p1<-FeaturePlot(cd4_seurat, features="PTGER2")+scale_color_gradient(low="grey80", high="red")
p2<-DimPlot(cd4_seurat,label=TRUE)+NoLegend()
p3<- VlnPlot(cd4_seurat, features="PTGER2")
p1+p2+p3
ggsave(paste0(figure_dir, "CD4_PTGER2.expression.png"),width=6, height=15)

## by cluster-based annotation (04_ManualProgramValidation.Rmd)
cowplot::plot_grid(plotlist=lapply(c("PTGER2", "SELPLG"), function(x){
  VlnPlot(cd4_seurat,group.by = "seurat_clusters_annot", features=x, sort=FALSE)+NoLegend()+
    scale_fill_manual(values=tcell_colors)
}))
ggsave(paste0(figure_dir, "Vln_CD4_byClusterAnnotated_PTGER2.SELPLG.png"),width=8, height=5)

## by projection from full cohort (03_ProjectPrograms.Rmd)
cowplot::plot_grid(plotlist=lapply(c("PTGER2", "SELPLG"), function(x){
  VlnPlot(cd4_seurat,group.by = "Program_projected", features=x, sort=FALSE)+NoLegend()+
    scale_fill_manual(values=tcell_colors)
}))
ggsave(paste0(figure_dir, "Vln_CD4_byProjectedProgram_PTGER2.SELPLG.png"),width=8, height=5)

```





## Correlation between PTGER2 expression and programs
```{r}
## Add on PTGER2 expression to scores
ptger_expr<- as.data.frame(cd8_seurat@assays$RNA@data["PTGER2",])
colnames(ptger_expr)<- c("PTGER2")
sum(rownames(ptger_expr)!=rownames(scores))
scores$PTGER2<- ptger_expr$PTGER2


## Linear regression
programs<-unique(cd8_seurat$Program_projected)
all_lm<- lapply(programs, function(x){
  tmp<- scores[,c(x, "PTGER2")]
  colnames(tmp)<- c("Program", "PTGER2")
  res<- summary(lm(Program~PTGER2, tmp))
  res<- list(r2=res$r.squared, pvalue=res$coefficients["PTGER2", "Pr(>|t|)"] )
  return(res)
})
names(all_lm)<- programs

## Plot scores vs expression
all_plots<- lapply(unique(cd8_seurat$Program_projected), function(x){
  tmp<- scores[,c(x, "PTGER2")]
  colnames(tmp)<- c("Program", "PTGER2\nExpression")
  tmp$color<- x
  sig<- all_lm[[x]]
  ggplot(tmp, aes(x=Program, y=`PTGER2\nExpression`, color=color))+
    geom_point()+
    ggtitle(x,
            subtitle = paste0("R2: ", round(sig$r2,3), ", p=", round(sig$pvalue,3)))+
    theme_classic()+
    scale_color_manual(values=tcell_colors[names(tcell_colors)==x])+
    theme(legend.position = "none")+
    geom_smooth(method="lm", se=FALSE, color="black", linetype="dashed")
})
cowplot::plot_grid(plotlist=all_plots)
ggsave(paste0(figure_dir, "ProgramScore_cor_PTGER2.png"), width=9, height=9)
```

## Correlation between SELPLG expression and programs
```{r}
## Add on selplg expression to scores
selplg_expr<- as.data.frame(cd8_seurat@assays$RNA@data["SELPLG",])
colnames(selplg_expr)<- c("SELPLG")
sum(rownames(selplg_expr)!=rownames(scores))
scores$SELPLG<- selplg_expr$SELPLG


## Linear regression
programs<-unique(cd8_seurat$Program_projected)
all_lm<- lapply(programs, function(x){
  tmp<- scores[,c(x, "SELPLG")]
  colnames(tmp)<- c("Program", "SELPLG")
  res<- summary(lm(Program~SELPLG, tmp))
  res<- list(r2=res$r.squared, pvalue=res$coefficients["SELPLG", "Pr(>|t|)"] )
  return(res)
})
names(all_lm)<- programs

## Plot scores vs expression
all_plots<- lapply(unique(cd8_seurat$Program_projected), function(x){
  tmp<- scores[,c(x, "SELPLG")]
  colnames(tmp)<- c("Program", "SELPLG\nExpression")
  tmp$color<- x
  sig<- all_lm[[x]]
  ggplot(tmp, aes(x=Program, y=`SELPLG\nExpression`, color=color))+
    geom_point()+
    ggtitle(x,
            subtitle = paste0("R2: ", round(sig$r2,3), ", p=", round(sig$pvalue,3)))+
    theme_classic()+
    scale_color_manual(values=tcell_colors[names(tcell_colors)==x])+
    theme(legend.position = "none")+
    geom_smooth(method="lm", se=FALSE, color="black", linetype="dashed")
})
cowplot::plot_grid(plotlist=all_plots)
ggsave(paste0(figure_dir, "ProgramScore_cor_SELPLG.png"), width=9, height=9)
```
